sorting harmony files
