+++ 1.6/Source/Mashed_Ashlands/Mashed_Ashlands/Harmony/Patch_MentalBreakWorker_RunWild.cs
@@ -1,29 +1,10 @@
 ﻿using HarmonyLib;
+using Verse.AI;
 using Verse;
 using RimWorld;
-using Verse.AI;
 
 namespace Mashed_Ashlands
 {
-    /// <summary>
-    /// Disables meditation outside during an ash storm
-    /// </summary>
-    [HarmonyPatch(typeof(MeditationUtility))]
-    [HarmonyPatch("SafeEnvironmentalConditions")]
-    public static class MeditationUtility_SafeEnvironmentalConditions_Patch
-    {
-        [HarmonyPostfix]
-        public static void Mashed_Ashlands_SafeEnvironmentalConditions_Patch(IntVec3 cell, Map map, ref bool __result)
-        {
-            if (__result)
-            {
-                if (!cell.Roofed(map))
-                {
-                    __result = !Utility.MapHasUnsafeCondition(map);
-                }
-            }
-        }
-    }
 
     /// <summary>
     /// Prevents the run wild mental break during an ash storm
