+++ 1.6/Source/Mashed_Ashlands/Mashed_Ashlands/Harmony/Patches_CliffRacer.cs
